+++ src/app/shared/module/poe/component/item-frame-properties/item-frame-properties.component.ts
@@ -24,6 +24,7 @@ export class ItemFramePropertiesComponent implements OnInit {
     this.properties = this.properties || [
       'weaponCriticalStrikeChance',
       'weaponAttacksPerSecond',
+      'weaponRange',
       'shieldBlockChance',
       'armourArmour',
       'armourEvasionRating',